Refs #45 Update call time of the sticky table headers initialisation.
It turns out that Angular will not give you a callback when a template is done rendering
by design. (See the discussion at [1][2])

As a mitigation strategy, we set a directive on the table which watches for assignment,
and then calls our jQuery plugin on "tables" at that time.

This WFM, but is not guaranteed to work by upstream, and shouldn't be viewed as more than
a temporary solution.

[1] https://github.com/angular/angular.js/issues/1306#issuecomment-8200414
[2] https://groups.google.com/forum/?fromgroups=#!topic/angular/SCc45uVhTt8