Remove windows-runner.py
This script is several years old, would no longer work, and refers to a hard coded path on a particular laptop that I don't believe we even have access to any more.

The general principle is sound for the case where for instance, one might like to deploy to a single laptop to conduct say, single user clinical testing of an application, but this is definitely the wrong place for such a script to live, even if it did work.