Informatics enjoys giving UUIDs to concepts. Whatever turns you on.
This implementation borrows the data model from FHIR CodeableConcept and uses it to make Opal lookuplists compatible
with for instance, a SNOMED CT Concept + synonyms.

This requries altering our serialised lookuplist datastructure and will require a migration to enter a  field on
all lookuplist instances in all applications, plugins and Opal itself.