Attempt to redirect the user to where they wanted to go on CSRF failure
This addresses the case where an unauthenticated user ends up with two
browser windows/tabs pointing at the login form.  Both versions of the
page will have valid CSRF tokens in their forms.  The user then logs in
with one of those pages and Django cycles the CSRF token, making the
second page's token invalid.  The user then tries to log in via the
second page and gets a CSRF failure/403.

This change redirects the user to their intended page via the `next`
param (falling back to /).  An unauthenticated user should be redirected
to the login page again (getting a valid CSRF token).